Move marker resolution from SSE to the "common" editor plugin
Those marker resolutions can be used by any editor, so we move them
to a more common plan for easier reuse.
