+++ org.eclipse.m2e.core.ui/META-INF/MANIFEST.MF
@@ -13,7 +13,7 @@ Export-Package: org.eclipse.m2e.core.ui.internal;x-internal:=true,
  org.eclipse.m2e.core.ui.internal.dialogs;x-friends:="org.eclipse.m2e.editor",
  org.eclipse.m2e.core.ui.internal.editing;x-friends:="org.eclipse.m2e.editor.xml,org.eclipse.m2e.editor,org.eclipse.m2e.refactoring",
  org.eclipse.m2e.core.ui.internal.lifecyclemapping;x-internal:=true,
- org.eclipse.m2e.core.ui.internal.markers;x-internal:=true,
+ org.eclipse.m2e.core.ui.internal.markers;x-friends="org.eclipse.m2e.editor",
  org.eclipse.m2e.core.ui.internal.preferences;x-internal:=true,
  org.eclipse.m2e.core.ui.internal.search.util;x-friends:="org.eclipse.m2e.editor",
  org.eclipse.m2e.core.ui.internal.util;x-friends:="org.eclipse.m2e.editor",
