+++ org.eclipse.m2e.editor.xml/META-INF/MANIFEST.MF
@@ -30,8 +30,6 @@ Require-Bundle: org.eclipse.core.runtime,
 Export-Package: org.eclipse.m2e.editor.xml;x-internal:=true,
  org.eclipse.m2e.editor.xml.internal;x-internal:=true,
  org.eclipse.m2e.editor.xml.mojo;x-internal:=true,
- org.eclipse.m2e.editor.xml.internal.lifecycle;x-internal:=true,
- org.eclipse.m2e.editor.xml.internal.markers;x-internal:=true,
  org.eclipse.m2e.editor.xml.preferences;x-internal:=true
 Bundle-ClassPath: .
 Bundle-Vendor: %Bundle-Vendor
