+++ org.eclipse.m2e.editor.xml/plugin.xml
@@ -24,17 +24,6 @@
 		target="org.eclipse.m2e.core.pomFile" />
   </extension>
 	<extension
-		point="org.eclipse.ui.ide.markerResolution">
-		<markerResolutionGenerator
-			class="org.eclipse.m2e.editor.xml.internal.markers.MavenMarkerResolutionGenerator"
-			markerType="org.eclipse.m2e.core.maven2Problem.pomloading">
-		</markerResolutionGenerator>
-		<markerResolutionGenerator
-			class="org.eclipse.m2e.editor.xml.internal.markers.MavenMarkerResolutionGenerator"
-			markerType="org.eclipse.m2e.core.maven2Problem.lifecycleMapping">
-		</markerResolutionGenerator>
-	</extension>
-	<extension
 		point="org.eclipse.core.filebuffers.annotationModelCreation">
 		<factory
 			contentTypeId="org.eclipse.m2e.core.pomFile"
