+++ org.eclipse.m2e.editor/plugin.xml
@@ -55,4 +55,16 @@
          </enablement>
       </objectContribution>
    </extension>
+
+	<extension
+		point="org.eclipse.ui.ide.markerResolution">
+		<markerResolutionGenerator
+			class="org.eclipse.m2e.editor.internal.markers.MavenMarkerResolutionGenerator"
+			markerType="org.eclipse.m2e.core.maven2Problem.pomloading">
+		</markerResolutionGenerator>
+		<markerResolutionGenerator
+			class="org.eclipse.m2e.editor.internal.markers.MavenMarkerResolutionGenerator"
+			markerType="org.eclipse.m2e.core.maven2Problem.lifecycleMapping">
+		</markerResolutionGenerator>
+	</extension>
 </plugin>
