+++ org.eclipse.m2e.editor/src/org/eclipse/m2e/editor/internal/markers/MarkerResolutionWrapper.java
@@ -12,7 +12,7 @@
  *      Anton Tanasenko - Refactor marker resolutions and quick fixes (Bug #484359)
  *******************************************************************************/
 
-package org.eclipse.m2e.editor.xml.internal.markers;
+package org.eclipse.m2e.editor.internal.markers;
 
 import org.eclipse.core.resources.IMarker;
 import org.eclipse.jface.text.IDocument;
