+++ org.eclipse.m2e.editor/src/org/eclipse/m2e/editor/internal/messages.properties
@@ -156,3 +156,15 @@ PomHyperlinkDetector_open_module=Open module project pom.xml at {0}
 PomHyperlinkDetector_open_property=Open definition of property {0}
 PomHyperlinkDetector_error_message=Can't open editor for {0}\n{1}
 PomHyperlinkDetector_error_title=Open Maven POM
+PomQuickAssistProcessor_name=Add Maven XML Schema declaration
+PomQuickAssistProcessor_remove_hint=It removes the current definition to rely on value inherited from parent
+PomQuickAssistProcessor_title_groupId=Remove groupId declaration
+PomQuickAssistProcessor_title_version=Remove version declaration
+MavenMarkerResolution_openManaged_label=Open declaration of managed version
+MavenMarkerResolution_openManaged_description=Open {0}
+LifecycleMappingProposal_all_desc=<html>Goal: <b>{0}</b><br/>ExecutionId: <b>{1}</b><br/>Phase: <b>{2}</b><br/>Plugin: {3}<br/><br/>{4}</html>
+LifecycleMappingProposal_execute_desc=This quickfix generates a plugin configuration snippet recognized by the m2e integration during project configuration. It marks the given goal to be executed during the Eclipse build.
+LifecycleMappingProposal_execute_label=Execute goal {0} as part of Eclipse build
+LifecycleMappingProposal_ignore_desc=This quickfix generates a plugin configuration snippet recognized by the m2e integration during project configuration. It marks the given goal as ignored for the purposes of the Eclipse build.
+LifecycleMappingProposal_ignore_label=Mark goal {0} as ignored in pom.xml
+LifecycleMappingProposal_workspaceIgnore_label=Mark goal {0} as ignored in eclipse preferences
