+++ pom.xml
@@ -83,9 +83,9 @@
       <url>https://download.eclipse.org/cbi/updates/license</url>
     </repository>
     <repository>
-      <id>Orbit 2020-09</id>
+      <id>Orbit 2021-03</id>
       <layout>p2</layout>
-      <url>https://download.eclipse.org/tools/orbit/downloads/drops/R20200831200620/repository</url>
+      <url>https://download.eclipse.org/tools/orbit/downloads/drops/S20210216215844/repository/</url>
     </repository>
     <repository>
     <id>Wild Web Developer</id>
@@ -202,6 +202,23 @@
         <artifactId>target-platform-configuration</artifactId>
         <configuration>
           <pomDependencies>consider</pomDependencies>
+          <executionEnvironment>JavaSE-11</executionEnvironment>
+          <filters>
+            <filter>
+              <!-- Remove as it may come as an undesired transitive dep from p2 under Java 11 -->
+              <type>p2-installable-unit</type>
+              <id>javax.xml</id>
+              <removeAll />
+            </filter>
+            <filter>
+              <!-- See bug https://bugs.eclipse.org/bugs/show_bug.cgi?id=566183 -->
+              <type>eclipse-plugin</type>
+              <id>org.apache.xerces</id>
+              <restrictTo>
+                 <versionRange>[2.12.1,3.0.0)</versionRange>
+              </restrictTo>
+            </filter>
+          </filters>
         </configuration>
       </plugin>
       <plugin>
